Add codes to script to control reporting of problems in the build

+++ OracleWebLogic/samples/12213-domain-home-in-image-wdt/build-archive.sh
@@ -4,9 +4,26 @@
 #
 #Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
 #
-script="${BASH_SOURCE[0]}"
-scriptDir="$( cd "$( dirname "${script}" )" && pwd )"
+# This script requires the following environment variables:
+#
+# JAVA_HOME            - The location of the JDK to use.  The caller must set
+#                        this variable to a valid Java 8 (or later) JDK.
+#
+
+if [ -z ${JAVA_HOME} ] || [ ! -e ${JAVA_HOME}/bin/jar ]; then 
+   echo "JAVA_HOME must be set to version of a java JDK 1.8 or greater"
+   exit 1
+fi
+echo JAVA_HOME=${JAVA_HOME}
+
+scriptDir="$( cd "$( dirname $0 )" && pwd )"
+if [ ! -d ${scriptDir} ]; then
+    echo "Unable to determine the sample directory where the application is found"
+    echo "Using shell /bin/sh to determine and found ${scriptDir}"
+    exit 1
+fi
+
 rm -Rf ${scriptDir}/archive
 mkdir -p ${scriptDir}/archive/wlsdeploy/applications
-jar cvf ${scriptDir}/archive/wlsdeploy/applications/simple-app.war -C ${scriptDir} simple-app/* 
-jar cvf ${scriptDir}/archive.zip  -C ${scriptDir}/archive wlsdeploy
+${JAVA_HOME}/bin/jar cvf ${scriptDir}/archive/wlsdeploy/applications/simple-app.war -C ${scriptDir} simple-app/* 
+${JAVA_HOME}/bin/jar cvf ${scriptDir}/archive.zip  -C ${scriptDir}/archive wlsdeploy
